requirements.txt
dotenv
redis==3.5.3
pytest==4.5.0
PyMySQL==0.9.3
allure-pytest==2.8.6
requests==2.20.0
PyYAML==5.3.1
python-dotenv=0.17.0